Create 1stSemesterControlBot.yml
This is the control bot. It would be connected to the computer and receive inputs from said computer. It would then translate the inputs  into signals and wirelessly send them to the main bot. The main bot would translate the signals back into inputs, which would then activate certain programs, allowing the main bot to move in multiple directions.